// taurus-protect-sdk-typescript/src/internal/openapi/apis/BlockchainApi.ts
export interface BlockchainServiceGetBlockchainsRequest {
    blockchain?: string;
    network?: string;
    includeBlockHeight?: boolean;
}

/**
 * BlockchainApi - interface
 * 
 * @export
 * @interface BlockchainApiInterface
 */
export interface BlockchainApiInterface {
    /**
     * This endpoint returns all the enabled blockchains
     * @summary Get blockchain entities
     * @param {string} [blockchain] 
     * @param {string} [network] 
     * @param {boolean} [includeBlockHeight] Include the current block height. Requires blockchain/network need to be specified.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof BlockchainApiInterface
     */
    blockchainServiceGetBlockchainsRaw(requestParameters: BlockchainServiceGetBlockchainsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetBlockchainsReply>>;

    /**
     * This endpoint returns all the enabled blockchains
     * Get blockchain entities
     */
    blockchainServiceGetBlockchains(requestParameters: BlockchainServiceGetBlockchainsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetBlockchainsReply>;

    /**
     * 
     * @summary Use /api/rest/v1/blockchains instead
     * @param {*} [options] Override http request option.
     * @deprecated
     * @throws {RequiredError}
     * @memberof BlockchainApiInterface
     */
    blockchainServiceGetBlockchainsInfoRaw(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetBlockchainsInfoReply>>;

    /**
     * Use /api/rest/v1/blockchains instead
     * @deprecated
     */
    blockchainServiceGetBlockchainsInfo(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetBlockchainsInfoReply>;

}

/**
 * 
 */
export class BlockchainApi extends runtime.BaseAPI implements BlockchainApiInterface {

    /**
     * This endpoint returns all the enabled blockchains
     * Get blockchain entities
     */
    async blockchainServiceGetBlockchainsRaw(requestParameters: BlockchainServiceGetBlockchainsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetBlockchainsReply>> {
        const queryParameters: any = {};

        if (requestParameters['blockchain'] != null) {
            queryParameters['blockchain'] = requestParameters['blockchain'];
        }

        if (requestParameters['network'] != null) {
            queryParameters['network'] = requestParameters['network'];
        }

        if (requestParameters['includeBlockHeight'] != null) {
            queryParameters['includeBlockHeight'] = requestParameters['includeBlockHeight'];
        }

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["Authorization"] = await this.configuration.apiKey("Authorization"); // ApiKeyTPV1 authentication
        }

        const response = await this.request({
            path: `/api/rest/v1/blockchains`,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => TgvalidatordGetBlockchainsReplyFromJSON(jsonValue));
    }

    /**
     * This endpoint returns all the enabled blockchains
     * Get blockchain entities
     */
    async blockchainServiceGetBlockchains(requestParameters: BlockchainServiceGetBlockchainsRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetBlockchainsReply> {
        const response = await this.blockchainServiceGetBlockchainsRaw(requestParameters, initOverrides);
        return await response.value();
    }

    /**
     * Use /api/rest/v1/blockchains instead
     * @deprecated
     */
    async blockchainServiceGetBlockchainsInfoRaw(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetBlockchainsInfoReply>> {
        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["Authorization"] = await this.configuration.apiKey("Authorization"); // ApiKeyTPV1 authentication
        }

        const response = await this.request({
            path: `/api/rest/v1/blockchains/info`,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => TgvalidatordGetBlockchainsInfoReplyFromJSON(jsonValue));
    }

    /**
     * Use /api/rest/v1/blockchains instead
     * @deprecated
     */
    async blockchainServiceGetBlockchainsInfo(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetBlockchainsInfoReply> {
        const response = await this.blockchainServiceGetBlockchainsInfoRaw(initOverrides);
        return await response.value();
    }

}